move EncodingType to azkaban.db module
